+++ code/src/nuvla/job/util.py
@@ -8,7 +8,6 @@ import random
 import logging
 import traceback
 import warnings
-from datetime import datetime
 
 PY2 = sys.version_info[0] == 2
 PY3_10 = (3, 10)
@@ -58,10 +57,6 @@ def retry_kazoo_queue_op(queue, function_name):
         random_wait(0.1, 5)
         logging.warning(
             'retry_kazoo_queue_op: Retrying {} on {}.'.format(function_name, queue.get()))
-
-
-def parse_cimi_date(date):
-    return datetime.fromisoformat(date[:-1] + '+00:00')
 
 
 def status_message_from_exception():